reorganized caches to be singletons
