+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/jita/conf/Configuration.java
@@ -38,7 +38,7 @@ public class Configuration implements Serializable {
 
     @Getter private AllocationStatus firstMemory = AllocationStatus.DEVICE;
 
-    @Getter private boolean debug = false;
+    @Getter private boolean debug = true;
 
     @Getter private boolean verbose = false;
