+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/jita/constant/CudaConstantHandler.java
@@ -20,10 +20,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.Semaphore;
 import java.util.concurrent.atomic.AtomicLong;
 
@@ -34,200 +32,30 @@ import java.util.concurrent.atomic.AtomicLong;
  */
 public class CudaConstantHandler extends BasicConstantHandler {
     private static Logger logger = LoggerFactory.getLogger(CudaConstantHandler.class);
-    protected Map<Integer, AtomicLong> constantOffsets = new HashMap<>();
-    protected Map<Integer, Semaphore> deviceLocks = new ConcurrentHashMap<>();
 
-    protected Map<Integer, Map<ArrayDescriptor, DataBuffer>> buffersCache = new HashMap<>();
-    protected Map<Integer, Pointer> deviceAddresses = new HashMap<>();
-    private Configuration configuration = CudaEnvironment.getInstance().getConfiguration();
-    protected NativeOps nativeOps = NativeOpsHolder.getInstance().getDeviceNativeOps();
-    protected FlowController flowController;
-
-    protected List<DataBuffer> protector = new CopyOnWriteArrayList<>();
-    private static final int MAX_CONSTANT_LENGTH = 49152;
-    private static final int MAX_BUFFER_LENGTH = 272;
-
-    protected Semaphore lock = new Semaphore(1);
+    protected static final ConstantHandler wrappedHandler = ProtectedCudaConstantHandler.getInstance();
 
     public CudaConstantHandler() {
 
     }
 
-    /**
-     * This method moves specified dataBuffer to CUDA constant memory space.
-     *
-     * PLEASE NOTE: CUDA constant memory is limited to 48KB per device.
-     *
-     * @param dataBuffer
-     * @return
-     */
     @Override
     public long moveToConstantSpace(DataBuffer dataBuffer) {
-        // now, we move things to constant memory
-        Integer deviceId = AtomicAllocator.getInstance().getDeviceId();
-        ensureMaps(deviceId);
-
-        AllocationPoint point = AtomicAllocator.getInstance().getAllocationPoint(dataBuffer);
-
-        long requiredMemoryBytes = AllocationUtils.getRequiredMemory(point.getShape());
-        // and release device memory :)
-
-        long currentOffset = constantOffsets.get(deviceId).get();
-        CudaContext context = (CudaContext) AtomicAllocator.getInstance().getDeviceContext().getContext();
-        if (currentOffset + requiredMemoryBytes >= MAX_CONSTANT_LENGTH || requiredMemoryBytes > MAX_BUFFER_LENGTH)  {
-            nativeOps.memcpyAsync(point.getPointers().getDevicePointer(), point.getPointers().getHostPointer(), requiredMemoryBytes, 1, context.getSpecialStream());
-            flowController.commitTransfer(context.getSpecialStream());
-
-            point.setConstant(true);
-            point.tickDeviceWrite();
-            point.tickHostRead();
-            return 0;
-        }
-
-        currentOffset = constantOffsets.get(deviceId).getAndAdd(requiredMemoryBytes);
-        if (currentOffset >= MAX_CONSTANT_LENGTH)  {
-            nativeOps.memcpyAsync(point.getPointers().getDevicePointer(), point.getPointers().getHostPointer(), requiredMemoryBytes, 1, context.getSpecialStream());
-            flowController.commitTransfer(context.getSpecialStream());
-
-            point.setConstant(true);
-            point.tickDeviceWrite();
-            point.tickHostRead();
-            return 0;
-        }
-
-
-        nativeOps.memcpyConstantAsync(currentOffset, point.getPointers().getHostPointer(), requiredMemoryBytes, 1, context.getSpecialStream());
-        flowController.commitTransfer(context.getSpecialStream());
-
-        long cAddr = deviceAddresses.get(deviceId).address() + currentOffset;
-        point.getPointers().setDevicePointer(new CudaPointer(cAddr));
-        point.setConstant(true);
-        point.tickDeviceWrite();
-        point.tickHostRead();
-
-        protector.add(dataBuffer);
-
-        return cAddr;
+        return wrappedHandler.moveToConstantSpace(dataBuffer);
     }
 
-    private void ensureMaps(Integer deviceId) {
-        if (!buffersCache.containsKey(deviceId)) {
-            if (flowController == null)
-                flowController = AtomicAllocator.getInstance().getFlowController();
-
-            try {
-                lock.acquire();
-                if (!buffersCache.containsKey(deviceId)) {
-                    buffersCache.put(deviceId, new ConcurrentHashMap<ArrayDescriptor, DataBuffer>());
-                    constantOffsets.put(deviceId, new AtomicLong(0));
-                    deviceLocks.put(deviceId, new Semaphore(1));
-
-                    Pointer cAddr = nativeOps.getConstantSpace();
-         //           logger.info("Got constant address: [{}]", cAddr);
-
-                    deviceAddresses.put(deviceId, cAddr);
-                }
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            } finally {
-                lock.release();
-            }
-        }
-    }
-
-    /**
-     * This method returns DataBuffer with contant equal to input array.
-     *
-     * PLEASE NOTE: This method assumes that you'll never ever change values within result DataBuffer
-     *
-     * @param array
-     * @return
-     */
     @Override
     public DataBuffer getConstantBuffer(int[] array) {
-      //  logger.info("getConstantBuffer(int[]) called");
-        ArrayDescriptor descriptor = new ArrayDescriptor(array);
-
-        Integer deviceId = AtomicAllocator.getInstance().getDeviceId();
-
-        ensureMaps(deviceId);
-
-        if (!buffersCache.get(deviceId).containsKey(descriptor)) {
-            // we create new databuffer
-        //    logger.info("Creating new constant buffer...");
-            DataBuffer buffer = Nd4j.createBuffer(array);
-
-            // now we move data to constant memory, and keep happy
-            moveToConstantSpace(buffer);
-
-            buffersCache.get(deviceId).put(descriptor, buffer);
-            return buffer;
-        } //else logger.info("Reusing constant buffer...");
-
-        return buffersCache.get(deviceId).get(descriptor);
+        return wrappedHandler.getConstantBuffer(array);
     }
 
-    /**
-     * This method returns DataBuffer with contant equal to input array.
-     *
-     * PLEASE NOTE: This method assumes that you'll never ever change values within result DataBuffer
-     *
-     * @param array
-     * @return
-     */
     @Override
     public DataBuffer getConstantBuffer(float[] array) {
-     //   logger.info("getConstantBuffer(float[]) called");
-        ArrayDescriptor descriptor = new ArrayDescriptor(array);
-
-        Integer deviceId = AtomicAllocator.getInstance().getDeviceId();
-
-        ensureMaps(deviceId);
-
-        if (!buffersCache.get(deviceId).containsKey(descriptor)) {
-            // we create new databuffer
-       //     logger.info("Creating new constant buffer...");
-            DataBuffer buffer = Nd4j.createBuffer(array);
-
-            // now we move data to constant memory, and keep happy
-            moveToConstantSpace(buffer);
-
-            buffersCache.get(deviceId).put(descriptor, buffer);
-            return buffer;
-        } // else logger.info("Reusing constant buffer...");
-
-        return buffersCache.get(deviceId).get(descriptor);
+        return wrappedHandler.getConstantBuffer(array);
     }
 
-    /**
-     * This method returns DataBuffer with contant equal to input array.
-     *
-     * PLEASE NOTE: This method assumes that you'll never ever change values within result DataBuffer
-     *
-     * @param array
-     * @return
-     */
     @Override
     public DataBuffer getConstantBuffer(double[] array) {
-//        logger.info("getConstantBuffer(double[]) called");
-        ArrayDescriptor descriptor = new ArrayDescriptor(array);
-
-        Integer deviceId = AtomicAllocator.getInstance().getDeviceId();
-
-        ensureMaps(deviceId);
-
-        if (!buffersCache.get(deviceId).containsKey(descriptor)) {
-            // we create new databuffer
-            //logger.info("Creating new constant buffer...");
-            DataBuffer buffer = Nd4j.createBuffer(array);
-
-            // now we move data to constant memory, and keep happy
-            moveToConstantSpace(buffer);
-
-            buffersCache.get(deviceId).put(descriptor, buffer);
-            return buffer;
-        } //else logger.info("Reusing constant buffer...");
-
-        return buffersCache.get(deviceId).get(descriptor);
+        return wrappedHandler.getConstantBuffer(array);
     }
 }
